+++ app/controllers/events_controller.rb
@@ -41,7 +41,7 @@ class EventsController < ApplicationController
 
   def destroy
     @event.destroy
-    redirect_to root_path
+    redirect_to dashboard_path
   end
 
   private