+++ spec/system/heartbeat_spec.rb
@@ -5,6 +5,6 @@ require 'rails_helper'
 RSpec.describe 'Basic Heartbeat Test', type: :system do
   it 'has an h1 tag' do
     visit root_path
-    expect(page).to have_content('Welcome to Cardswipr')
+    expect(page).to have_content('Welcome to CardSwipr')
   end
 end